Refactor: move default model implementation from app to api
