+++ apps/openstacknode/api/src/main/java/org/onosproject/openstacknode/api/DefaultDpdkInterface.java
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.onosproject.openstacknode.impl;
+package org.onosproject.openstacknode.api;
 
 import com.google.common.base.MoreObjects;
-import org.onosproject.openstacknode.api.DpdkInterface;
 
 import java.util.Objects;
