+++ apps/openstacknode/api/src/main/java/org/onosproject/openstacknode/api/DefaultOpenstackSshAuth.java
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.onosproject.openstacknode.impl;
+package org.onosproject.openstacknode.api;
 
 import com.google.common.base.MoreObjects;
-import org.onosproject.openstacknode.api.OpenstackSshAuth;
 
 import java.util.Objects;
